requirements.txt
google-genai>=1.0.0
python-dotenv
pysubs2
